api/requirements.txt
fastapi[standard]==0.115.12
alembic==1.15.1
alibabacloud-sts20150401==1.1.5
beautifulsoup4==4.13.3
httpx==0.28.1
openai==1.70.0
pydantic==2.10.6
pymilvus[model]==2.5.6
PyMySQL==1.1.1
python-dotenv==1.1.0
redis==5.2.1
SQLAlchemy==2.0.40
chonkie==0.5.1
celery==5.5.0
tiktoken==0.9.0
torch==2.6.0
apscheduler==3.11.0
pyJWT==2.10.1
python-jose==3.4.0
opencv-python==4.11.0.86
magic-pdf==1.3.7
ultralytics==8.3.106
doclayout_yolo==0.0.3
rapid_table==1.0.5
ftfy==6.3.1
shapely==2.1.0
pyclipper==1.3.0.post6
omegaconf==2.3.0
playwright==1.51.0